Move update button to the outside
